Rename org.enso.logging.service.remote to org.enso.logging.service.common
