+++ lib/java/logging-service-common/src/main/java/org/enso/logging/service/common/LogJob.java
@@ -1,4 +1,4 @@
-package org.enso.logging.service.remote;
+package org.enso.logging.service.common;
 
 import java.util.concurrent.CompletableFuture;
