+++ lib/java/logging-service-opensearch/src/main/java/module-info.java
@@ -5,7 +5,7 @@ module org.enso.logging.service.opensearch {
   requires ch.qos.logback.core;
   requires static org.enso.logging.service.logback;
   requires org.enso.logging.service;
-  requires org.enso.logging.service.remote;
+  requires org.enso.logging.service.common;
   // For jsoniter_scala
   requires org.enso.scala.wrapper;
   requires org.slf4j;
