+++ lib/java/logging-service-opensearch/src/main/java/org/enso/logging/service/opensearch/OpenSearchLogFormatter.java
@@ -4,8 +4,8 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.regex.Matcher;
-import org.enso.logging.service.remote.ApiMessage;
-import org.enso.logging.service.remote.LogMessage;
+import org.enso.logging.service.common.ApiMessage;
+import org.enso.logging.service.common.LogMessage;
 
 public class OpenSearchLogFormatter {
