+++ lib/java/logging-service-opensearch/src/test/java/org/enso/logging/service/opensearch/TestLogFormatter.java
@@ -11,8 +11,8 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.List;
 import java.util.Map;
-import org.enso.logging.service.remote.ApiMessage;
-import org.enso.logging.service.remote.LogMessage;
+import org.enso.logging.service.common.ApiMessage;
+import org.enso.logging.service.common.LogMessage;
 import org.junit.Assert;
 import org.junit.Test;
