+++ lib/java/logging-service-telemetry/src/test/java/org/enso/logging/service/telemetry/TestLogFormatter.java
@@ -12,8 +12,8 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.List;
 import java.util.Map;
-import org.enso.logging.service.remote.ApiMessage;
-import org.enso.logging.service.remote.LogMessage;
+import org.enso.logging.service.common.ApiMessage;
+import org.enso.logging.service.common.LogMessage;
 import org.junit.Assert;
 import org.junit.Test;
