+++ lib/java/logging-service-telemetry/src/test/java/org/enso/logging/service/telemetry/TestTelemetry.java
@@ -15,12 +15,12 @@ import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.IntStream;
-import org.enso.logging.service.remote.AuthenticationData;
-import org.enso.logging.service.remote.Credentials;
-import org.enso.logging.service.remote.LogJob;
-import org.enso.logging.service.remote.LogJobsProcessor;
-import org.enso.logging.service.remote.LogMessage;
-import org.enso.logging.service.remote.TokenRefresher;
+import org.enso.logging.service.common.AuthenticationData;
+import org.enso.logging.service.common.Credentials;
+import org.enso.logging.service.common.LogJob;
+import org.enso.logging.service.common.LogJobsProcessor;
+import org.enso.logging.service.common.LogMessage;
+import org.enso.logging.service.common.TokenRefresher;
 import org.enso.shttp.HybridHTTPServer;
 import org.enso.testkit.RetryTestRule;
 import org.junit.After;
