+++ lib/java/logging-service-telemetry/src/test/java/org/enso/logging/service/telemetry/TestTelemetryBrokenServer.java
@@ -14,11 +14,11 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
-import org.enso.logging.service.remote.AuthenticationData;
-import org.enso.logging.service.remote.LogJob;
-import org.enso.logging.service.remote.LogJobsProcessor;
-import org.enso.logging.service.remote.LogMessage;
-import org.enso.logging.service.remote.TokenRefresher;
+import org.enso.logging.service.common.AuthenticationData;
+import org.enso.logging.service.common.LogJob;
+import org.enso.logging.service.common.LogJobsProcessor;
+import org.enso.logging.service.common.LogMessage;
+import org.enso.logging.service.common.TokenRefresher;
 import org.enso.testkit.RetryTestRule;
 import org.junit.After;
 import org.junit.Before;
