+++ lib/java/logging-service-telemetry/src/test/java/org/enso/logging/service/telemetry/Utils.java
@@ -8,8 +8,8 @@ import java.time.format.DateTimeFormatter;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Executors;
-import org.enso.logging.service.remote.Credentials;
-import org.enso.logging.service.remote.LogJob;
+import org.enso.logging.service.common.Credentials;
+import org.enso.logging.service.common.LogJob;
 import org.enso.shttp.HTTPTestHelperServer;
 import org.enso.shttp.HybridHTTPServer;
 import org.enso.shttp.cloud_mock.CloudMockSetup;
